+++ apps/appointment-booking-N3/src/data/index.ts
@@ -2,6 +2,7 @@ export { doctors, searchDoctors, getDoctorById } from './doctors'
 export { getTimeSlots, getAvailableDates, getSlotsForDate } from './timeSlots'
 export { stores, searchStores, getStoreById } from './stores'
 export { cmsContent, getActiveCMSContent } from './cms'
+export { symptoms, specialties, getSpecialtyForSymptom } from './symptoms'
 export {
   simulateApiDelay,
   simulateApiError,
@@ -12,4 +13,6 @@ export {
   apiGetSlotsForDate,
   apiSearchStores,
   apiGetCMSContent,
+  apiFastLaneMatch,
 } from './api'
+export type { FastLaneRequest, FastLaneMatchResult } from './api'